Redo thread test
The entire thread previously create was inadequate due to its innacurate
timing, for example, slightly random number of threads occuring at
inaccurate time in total. The thread is now very accurate: it loops the
exact number of times specified and (depending on the machine used)
finishes in the total time expected by the microsecond.

Also, the entire thread is more simple due to removing certain
variables.